Update our security policy to include all releases in the current series
We don't want people to jump from the last release in the previous
series (4.2) straing to the first release in the current series that is
support (5.1), so we should be supporting the entire series for severe
security issues.

This will not be a problem because usually we only have 3 releases in a
series and backporting patches inside the same series is not harder than
backporting to a previous series.