Added --editor (-e) option to open generated files in the user's editor
This is a retry of #8553.

Added support for generators added later from the original PR, and extract
the function on the primary file into method.

Also avoied adding the `editor` option to `Generators::Base`.
If add it to `Generators::Base`, `editor` option will be output to the help
of generator which does not support it. This confuses the user.