Relation values into their own instance vars
Moves the various components of an AR relation into their own instance variables. This cuts down on the amount of time that has to be spent in `get_value` and `set_value` accessing a hash, versus dealing with the instance variables directly. It also has very slightly better memory performance by not keeping a hash around for each relation object.