Handle rollback of Postgres index when using CAST
When creating an index in Postgres `CAST(created_at AS date)` is converted to `((created_at)::date)`.
This difference causes the `index_name` method to return 'index_payments_on_CAST_created_at_AS_date' when the index is created with the CAST statement and 'index_payments_on_created_at_date' when the postgres shorthand syntax is passed in during removal.

By adding an additional check against the index name we are able to handle this scenario appropriately.

Fixes #32621