Add core extension Enumerable index_grouped_by.
Updates the core extension to include a new method for Enumerable called
index_grouped_by. The behaviour of this method is almost the same as
that of the index_by method. The difference is that index_grouped_by
makes lists of all hash values, and appends items with the same key to
the same list. This allows for indexing an enumerable by a key that is
not unique.