restores original intention in constants guide, reworded [ci skip]
Once all technicalities have been introduced, the purpose of this
paragraph is to precisely unroll expressions like "the String class".
That way, the reader is forced to clearly separate concepts that Ruby
programmers often have kind of blurred:

  * Constants are storage, like variables.

  * Ruby does not have syntax for class or module names.

  * `String` is a regular constant that holds a value. In this case, the
    value happens to be a class object.

  * Constants are stored in class and module objects. In the case of
    `String`, the holder is the class object stored in the `Object`
    constant.

Understanding that paragraph the way is written is important to
accomplish this objective.

References #32818.