Add option to ActiveStorage::Blob to set extract_content_type_from_io
This adds a boolean extract_content_type_from_io to ActiveStorage::Blob
methods #create_after_upload, #build_after_upload and #upload. It
allows a user to manually override the automatic MimeType extraction from
the io. This provides a patch for when Marcel::MimeType.for returns an
incorrect content type when #extract_content_type is called from the #upload method.