Consolidate errors when loading adapter gem fails into `LoadError`
If not using Bundler and the specified adapter gem does not exist,
`rubygems` raises `Gem::MissingSpecError`.

`Gem::MissingSpecError` inherits `LoadError`, but the argument of `initialize`
is different.
https://github.com/rubygems/rubygems/blob/9054079ce8a1ad05dd206a3562339b74712c7fc6/lib/rubygems/errors.rb#L27..L31

As a result, `ArugmentError` occurs when re-raising the exception.

Although it is possible to check the error class and specify appropriate
arguments for each of them, I think that it is sufficient to just raise
`LoadError`.

Together, when using Bundler, fixed that gem which error occurred could
not be checked correctly.

Fixes #32830.