Handle mutated values at assert_changes
Take the following assertion:

```ruby
assert_changes 'TestingPublisher.events', to: [event] do
  event.publish(time)
end
```

`TestingPublisher.events` is an array, that is mutated every time
`event.publish` is executed. The assertion always fails in this case,
because `assert_changes` cannot figure out that
`TestingPublisher.events` has not changed it's reference, but has been
mutated.

I'm not exactly sure, whether we should try to handle this case, as
testing the reference preservation is also useful, but here is a simple
fix that handles the example above well.