Handle pid collisions safely in ActiveRecord
Due to pid recycling, it is possible for a process to have the same pid as its
grandparent's. A map from pid to connection pools is kept in
`ActiveRecord::ConnectionAdapters::ConnectionHandler`. If a process has the
same pid as its grandparent, then we might fetch a connection pool from this
map that has already been discarded. This can lead to errors when
`#establish_connection` is called in a process whose pid is the same as its
grandparent's.

In addition to discarding the connection pools belonging to an ancestor
process, we should also delete them from the pid map to avoid the pid collision
danger in descendant processes.

Fixes #33022.