Allow fstring as a safe raw SQL string
#27947 was intended to address `Post.order(params[:order])` which
leading to SQL injection.

But the protection caused very annoying warnings, especially for calling
by a string literal which invoke a function like `Post.order("length(title)")`.

Calling by a string literal should be safe since it is not an user
input. So I'd like to allow string literals as a safe raw SQL string.

There is no reliable way to distinguish between `params[:order]` and
`"length(title)"`, but at least `params[:order]` as an user input is
always a mutable string, and `"length(title)"` will be a fstring if
`# frozen_string_literal: true`.

So I think that we can add fstring to the allowlist to mitigate the
annoying warnings without losing the original intention of #27947.

Resolves #32995.