Add ability to configure cache notifications info
This may seem like an unnecessary refactoring but some apps want / need
to configure the information passed to the query cache logger. In order
to do that we can add a method here that can be easily overridden by the
app itself, rather than hacking the query cache logger to include that
information.

To override apps can call

```
def cache_notifications_info
  super.merge(connected_host: "hostname")
end
```

This will take what's already in the query cache logger and add
`@something="yea"` to the object.

At GitHub we use this to log the number of queries that are cached, the
connection host and the connection url.