Specify lang attribute in generated app layout
This adds the lang attribute to the html tag in the generated
application.html.erb layout. The W3C [states][1]:

> Always use a language attribute on the html tag to declare the default
> language of the text in the page.

Rails exposes the code for the current language with the `locale`
helper. This seemed like the obviuous choice to set the language
properly and automatically.

Why is this a problem?

First, declaring what language a page is written in is a good web
practice.

Second, some browsers such as Google Chrome will ask if a site needs to
be translated if it detects there are many non-native words on the
page. It's possible to confuse this translation detector if, for
example, you have a quirky company name on a page with little other text
like a promo page. Adding the lang attribute puts a stop to this
unhelpful suggestion by the browser.

[1]:
https://www.w3.org/International/questions/qa-html-language-declarations