Clarify warning about `after_commit` for both create & update
The docs have contained a warning that
> Using both `after_create_commit` and `after_update_commit` in the same model will only allow the last callback defined to take effect, and will override all others.

to this I suggest updating it to 

> with the same method in the same model

This is because that is what is accurate.

For example, if I have

```ruby
  after_create_commit :say_apple
  after_create_commit :say_banana
  after_update_commit :say_cherry
  after_update_commit :say_dragonfruit
  
  def say_apple
    puts "apple"
  end

  def say_banana
    puts "banana"
  end

  def say_cherry
    puts "cherry"
  end

  def say_dragonfruit
    puts "dragonfruit"
  end
```
then creating outputs both `apple` and `banana` and updating outputs both `cherry` and `dragonfruit`. However, when the same method is used, such as in the example provided, that's when there is actually an override.