Rollback transaction if incorrect isolation level is passed
This fixes an issue where a transaction is left open if an incorrect
isolation level is passed as an option. eg:
`Table.transaction(isolation: "foobar") { #... }`
This behaviour can have the side effect of unknowingly leaving lock
entries in the database until another transaction is committed or rolled
back.

This is an issue only for postgresql databases since setting transaction
isolation levels must happen inside transactions. In mysql, however,
specifying isolation levels outside transactions will apply it to the
next transaction that is opened; so we don't need to change the mysql
adapter.

This solution rescues the `KeyError`, executes a transaction rollback, then
re-raises the `KeyError`, leaving behaviour the way it currently is,
except rolling back the failed transaction.