Don't show unneeded deprecation warning on server restart.
If booting a server via `rails s -u puma`, we'd convert the
option to a `using` positional.

When using `rails restart` our `restart_command` would the
option converted to the using positional and put that in
the restart command.

Thus we'd show users deprecation warnings for our own code.

Fix that by converting a passed positional to an option instead.

Also: fix initialize method signature. The local_options are an
array, not a hash. But don't even bother assigning defaults as
Thor passes them in.