Tell the user that the seeds transaction got rolled back
By doing this users will know for sure that they didn't leave the
database in an inconsistent state.

I moved the opening of the transaction from
"activerecord/lib/active_record/railties/databases.rake" to
"railties/lib/rails/engine.rb" because there already were tests for the seeds
file being loaded there, that I could base my tests on.

I suppose modifying `$stderr` is generally a bad idea, but it was the only way
I could think off to also verify that some output was given to the user. I
consider this output to be important and worth testing, so I decided it was a
worthy trade-off.