Consistent initializer for ActionDispatch::Http::UploadedFile
ActionDispatch::Http::UploadedFile.new accepts some parameters in
it's initializer for content-type and headers. These are named
`type` and `head`, which is unintuitive.  This PR allows the object
to be initialized with `content_type` and `headers` to bring
consistency with the name of the actual attributes. It also
maintains backwards compatibility.

    ActionDispatch::Http::UploadedFile.new({
      tempfile:     tempfile,
      content_type: 'image/svg+xml',
      headers:      'some_headers',
      filename:     'filename.svg',
    })