Conditionally use `helper_method` in Flash concern
I was attempting to use the `flash` functionality in a `Metal`
controller. When including the `flash` concern I received the following
error:

    NoMethodError: undefined method `helper_method'....

Either:

- `AbstractController::Helpers` should be a dependency of
  `ActionController::Flash`
- `ActionController::Flash` should not require the existence of
  `AbstractController::Helpers`.

Since my use case (set a flash and redirect) has no need for the helper
method and that is a common use case, making the dependency conditional
seemed the better option.

NOTE: This is similar to issue #21067 only the error is within Rails
itself while that issue had the error within Devise.