Clarify new_framework_defaults instructions.
The instructions for using new_framework_defaults is confusing, primarily
due to the word 'flip', which implies changing a boolean value. This isn't
how the new_framework_defaults is to be used. Instead, users are expected
to _uncomment_ each new default and try each out, one at a time.

Clarify the instructions to match the intended use of new_framework_defaults.