Improve handling of querying and loading data with `ignored_columns`
There are multiple situations that led to investigating how
`ignored_columns` works, how it interacts with querying and how to make
it better in more situations.

The first problem is that the use of `ignored_columns` in a class was
breaking migrations that ran after the removal of a column in that class
(a column that was not added to `ignored_columns`). This turned out to
be because when `ignored_columns` has entries, AR was exploding the full
list of known columns in the query (including the column that was just
removed , instead of letting `Arel.star` just return the database's known
list. While this is solvable by adding `reset_column_information` in the
migration between column removal and the failing code, this research led
to our next concern.

Our second concern is the performance effects of splatting out the full
column list for every query for a model with `ignored_columns` set. We
haven't done explicit performance testing but processing a list of
`arel_columns` is definitely slower than a simple `Arel.star`, and this
could easily add up for high traffic sites.

As such, we wanted to find a way to support `ignored_columns` properly,
including keeping the correct `reload` functionality, without resorting
to a full set of columns for every query.

This commit contains two changes to fix the above concerns.

* When a select set is explicitly given, remove the columns that are in
the `ignored_columns` list.
* When loading a record from the database, ignore the attributes that
are in the `ignored_columns` list.