Implement ActiveRecord::Relation#eql? and ActiveRecord::Relation#hash
Per the Ruby documentation regarding `eql?`:

    For objects of class Object, eql? is synonymous with ==. Subclasses
    normally continue this tradition by aliasing eql? to their
    overridden == method....

My reading is that unless there is another good reason not to alias
`eql?` to `==` it should be.

In addition, per the Ruby documentation regarding the `hash` method:

    must have the property that a.eql?(b) implies a.hash == b.hash

For this reason I have also implemented hash so that if two relation
objects contain the same records they return the same hash value.

My specific need is that I am doing a `group_by` where the key is a
`ActiveRecord::Relation`. Internally `group_by` is using a `Hash` which
a `Hash` uses `eql?` and `hash` to index.

The only relevant pre-existing issue I found was:

    https://github.com/rails/rails/issues/11947#issuecomment-22967919

In that issue it was offered that `eql?` could alias `==` if that was
needed by rspec, but since it wasn't the issue went stale.