Ensure references creates id before type
Prior to this commit, using `references` with `polymorphic: true` in a
migration created an id column before creating the  corresponding type
column:

https://github.com/rails/rails/commit/d26704a15f88d384dd282425daa832affdb5f8c1

After that commit, the order is reversed.

This PR restores the previous order and adds a test to prevent
regressions.

I'm not aware of any production impact for the column ordering change, but it
broke the schema linting on github/github.