Use string lengths instead of regexp to extract path
The regexp was introduced in 186ac4cdaa911a9af659a29f2179a19b99dea13b,
and looks cosmetic. While they should be functionally identical in
theory, in practice, case insensitive (but preserving) filesystems can
give results that are differently-cased from the pattern we supplied.

I don't know how to force the filesystem to do the surprising thing,
even when running in an environment that _could_, so no new test.