Add the ability to swap connections in Active Record
This is the first iteration of adding support for multiple databases in
Rails. Please note that if this version is accepted the next step will
be adding support for QueryCache on multiple handlers and allowing the
application to register multiple handlers.

This PR adds the ability for an application to have multiple handlers
and switch between those handlers in order to change connections.

By default Rails will register a `default` handler (for write
connections), and a `readonly` handler (for readonly connections). On
load these handlers are registered.

In an application model we can connect to the database using
`establish_connection` which will determine which handler to connect to
based on options passed in. For example given a configuration like this:

```
development:
  primary:
    database: my_primary_db
    user: root
  primary_replica:
    database: my_primary_db
    user: ro_user
    replica: true
  animals:
    database: my_animals_db
    user: root
  animals_replica
    database: my_animals_db
    user: ro_user
    replica: true
```

A model would connect like this:

```
class AnimalsBase < ApplicationRecord
  establish_connection :animals, :default
  establish_connection :animals, :readonly
end
```

This will ensure when a model under `AnimalsBase` is loaded, for example
`Dog`, `Dog` will have a connection to the write db and the readonly db
through the connection handlers.

To swap between databases applications can call `use_write_connection`
or `use_readonly_connection`. If no handler or connection type is
provided the application will default to the write db (later I'd like to
make this configurable, for example at GitHub we default to the readonly
connections).

Examples:

```
Animals.use_write_connection do
  # dog will be created
  Dog.create!
end

Animals.use_readonly_db do
  # exception will be thrown due to using a readonly user
  Dog.create!
end
```