Do not reset an already correct counter cache
In the best case, when all counter rows are already up to date (so
there's nothing to reset) this will avoid an unnecessary UPDATE query.
This drops the number of queries for each record from 3 (find, count, &
update) to two (find & count). With a large number of records, this can
be a big savings.

NOTE: if using the `touch: true` option, we'll still issue the UPDATE.
This is to preserve the current, though not specified, behavior of
`::reset_counters(id, association, touch: true)` wherein all records
would have their timestamp updated. I don't know if that's the "right"
behavior, but it is consistent.