Prevent deadlocks from touch callbacks on chained associations
Even with great care to arrange our model updates consistently within
transaction blocks, ActiveRecord could still cause deadlocks based on the
sequence of models needing to touch associations before commit.

The order that records are appended to the transaction callback array should be
sorted in a sequence that is:

1. Consistent between different transactions operating on the same tables, and
2. Least likely to conflict with the order of statements across different tables

Consider a simple hierarchy of Child belongs to Parent with touch: true. We can
easily create a situation where the deferred touches are listed out of order on
transaction records, based on different data conditions:

Transaction A

    Child.nice.each(&:send_gifts)
    # => [
    #   #<Child id: 1, parent_id: 1>,
    #   #<Child id: 2, parent_id: 2>
    # ]
    current_transaction.records
    # => [
    #   #<Parent id: 1>
    #   #<Parent id: 2>
    # ]

Transaction B

    Child.naughty.each(&:send_coal)
    # => [
    #   #<Child id: 3, parent_id: 2>,
    #   #<Child id: 4, parent_id: 1>
    # ]
    current_transaction.records
    # => [
    #   #<Parent id: 2>
    #   #<Parent id: 1>
    # ]

If these two transactions were to run concurrently, the `before_commit` callback
sequence could end up as:

|     | Transaction A             | Transaction B             |               |
| --- | ---                       | ---                       | ---           |
| 1   | UPDATE parents WHERE id=1 |                           | A locks row 1 |
| 2   |                           | UPDATE parents WHERE id=2 | B locks row 2 |
| 3   | UPDATE parents WHERE id=2 |                           | A waits for B |
| 4   |                           | UPDATE parents WHERE id=1 | B waits for A |

As each transaction is holding a row lock and waiting on the other to complete,
the database has no choice but to ROLLBACK one of these transactions with a
corresponding deadlock error.

To prevent this, we sort the transaction callback records in order of:

1. *Greatest depth* so when different associations reference a parent through
different paths, the deepest dependencies are applied first. This mostly mimics
the current implementation as associations get traversed upwards. (Polymorphic
associations are not counted for depth since they are indeterminate.)
2. *Table name* so different models that refer to the same table are applied
together in sequence, and not interleaved with other models.
3. *Primary key* as records from the same table can be listed for touch
callbacks in different order, they need to be updated in sequence.