Only recurse when there's more preloading to do
`preloaders_for_one` drops `nil`s in the incoming array, in
`grouped_records`. As a result, if it returns a loader with a non-empty
`preloaded_records` then `records` must have at least one non-nil
element.

`preloaders_on` returns an empty array if it's passed an empty array and
does not modify any other state, so it only needs to be called if
preloaded records are present.

Taken together: it's only necessary to calculate `polymorphic_parent` if
`recs` is non-empty, and then is always possible to do so since `records`
must have at least one non-`nil` element.