Documentation: Using Models in Your Migrations
Rails Guides till version 4.1, had a documentation on how to safely use Models
in migration files. Such documentation should be preserved as they help in
knowing what is the recommended way of solving a particular problem.

To quote Rails guide @ version 4.1 ( [link](https://guides.rubyonrails.org/v4.1/migrations.html#using-models-in-your-migrations) )

When creating or updating data in a migration it is often tempting to use one
of your models. After all, they exist to provide easy access to the underlying
data. This can be done, but some caution should be observed.