1.8x Faster Partial Caching - Faster Cache Keys
In rails/rails#35257 Aaron pointed out that partial caching is very slow when the partial is very
 small. When I ran the same benchmark locally (without this patch) I get around 5.97x slower than
no caching. With this PR I get 3.29x slower. This represents a 1.8x improvement in speed.

With this PR i'm seeing a 1.09x faster performance improvement in CodeTriage response time
via derailed_benchmarks.

Currently, cache keys are generated via method calls. When a single object is retrieved the raw object that the cache key is built from must be iterated over twice. Once to build the cache key and another for the cache version. In addition to generating the cache key for interacting with the cache, it is also re-generated for logging inside of the `instrument` call. These two things make generating cache keys slow.

To make this faster I moved the key generation to an object. As elements are added to the key object a cache key and cache version are built incrementally. Since we can build both at the same time we can optimize this execution. In this model we only generate cache key once, retrieving the cache key multiple times on this object is now fast.

One other cause of slowness is that when the partial is being rendered the view generates and prepends the "combined_fragment_cache_key" to each of the cache keys. Because our cache objects are now composable we can instead generate the "combined_fragment_cache_key" once and re-use the same object for generating N cache keys.

This is building off the work of rails/rails#33944. The main difference here is that this PR is backwards compatible.

You can run this benchmark by following these instructions:

```
$ git clone https://github.com/codetriage/codetriage
$ cd codetriage
$ git checkout schneems/keymaster-reproduction
$ bundle install
$ bundle exec ruby perf_scripts/partial_cache.rb
-- create_table(:customers, {:force=>true})
   -> 0.0092s
/Users/rschneeman/Documents/projects/codetriage/perf_scripts/partial_cache
<#ActiveSupport::Cache::MemoryStore entries=1000, size=352893, options={}>
Warming up --------------------------------------
collection render: no cache
                        27.000  i/100ms
collection render: with cache
                         7.000  i/100ms
Calculating -------------------------------------
collection render: no cache
                        272.792  (± 2.9%) i/s -      1.377k in   5.052080s
collection render: with cache
                         82.970  (± 2.4%) i/s -    420.000  in   5.066222s

Comparison:
collection render: no cache:      272.8 i/s
collection render: with cache:       83.0 i/s - 3.29x  slower
```