ActiveSupport::Cache::Key object - Faster Cache Keys
This PR moves the existing logic of `expanded_key` and `expanded_version` methods from ActiveSupport::Cache to their own class. This PR is a performance refactor.

Currently, when a cache entry is fetched the method, `normalize_key` (which calls `ecpanded_key`) is called on the key multiple times both for generating the actual key and for use with the cache store and logging. The existing version of this method is quite expensive and allocates memory. The biggest example is that it generates new arrays when the input is an array by calling `compact` and then again by calling `Array#to_param`. By moving the required logic to generate a key to a class, the need to generate intermediate array objects is drastically reduced. The key is generated as elements are added to the object as opposed to when they are needed. This behavior alleviates the need to re-generate the same key value for cache retrieval and logging.

One crucial property of this new cache class is that keys are composable. This composability means that if you pass the same objects into a class instance in the same order, it will produce the same key. This property allows us to optimize cache key generation in `collection_caching.rb` as the root of all elements being cached are the same.

## Performance

### Memory

Before:

```
Total allocated: 741750 bytes (6645 objects)
Total retained:  18917 bytes (166 objects)
```

After:

```
Total allocated: 707548 bytes (6006 objects)
Total retained:  18917 bytes (166 objects)
```

Diff:

```
(741750 - 707548) / 741750 # => ~4.6 %
```

### Performance

Here's the raw results: https://gist.github.com/schneems/7e808e7a866c1bc3a2d50b5e0ae4db88 for 63 runs, each with 1000 hits to the index of CodeTriage.

On average this patch is **7.24%** faster than the commit before it. This number is statistically significant.