Decrease memory allocations in cache.rb
The `merged_options` method is private and the output is never mutated.
Using this info we can get rid of the `dup` behavior. We can also
eliminate the `merge` call in the case where all the options being merged are the same.

Returned results are frozen as an extra layer of protection against
mutation.

Before

```
Total allocated: 741749 bytes (6642 objects)
```

After

```
Total allocated: 734039 bytes (6648 objects)
```

Diff

```
(741749 - 734039) / 741749.0 => ~ 1.0 %
```

If you don't feel comfortable modifying this method, we could rename it
and only use it internally.