Refactored abstract MySQL adapter to support lazy version check.
Will allow sub classes to override the protected
`#check_version` method hook if desired.

For example, this will be most helpful in sub classes that wish
to support lazy initialization because the version check can
be postponed until the connection is ready to be initialized.

This is a backport of what is being applied in Rails `master`.