[ActiveStorage] Enable withCredentials for authenticated cross-origin requests
We have a VueJS fontend on one domain which include the ActiveStorage npm package to do uploads on another backend domain. To make it work we had to enable withCredentials in the blob_record.js file, otherwise the cookies are not passed on the backend. 
I've seen some issues around this topic but still not a pull-request to fix this specific simple issue.
To me seems ok to have withCredentials always enabled, what do you think?