Default to created_at ordering when using UUID as PK
Although Rails supports UUID as a primary key, ordering by id, and therefore .first and .last, are no longer meaningful. The closest way to replicate an id ordering would be to utilize created_at when available.

The two other methods often suggested are:

1) Create a default_scope on created_at. Default scopes should be avoided in general and will cause issues with many joins/groupings.
2) Override first and last at the model level. This does not replicate the same functionality of the first and last methods (limit, etc)

Outside of explicitly calling order each time you use .first and .last or monkey patching Active Record, there don't seem to be good solutions available.

Since the Rails default is to order by ID, we can check to see whether the primary_key is of type :uuid, and use created_at to order if it's available.