Default to created_at ordering when using UUID as PK
### Summary

Although Rails supports UUID as a primary key, ordering by id, and therefore .first and .last, are no longer meaningful. The closest way to replicate the original order by ID would be to utilize created_at when available.

Since the Rails default is to order by ID, we can check to see whether the primary_key is of type :uuid, and use created_at to order if it's available.

This is not meant to augment the default Rail's behavior, but rather restore it.

### Other Information

The two other methods often suggested are:

Create a default_scope on created_at. Default scopes should be avoided in general and will cause issues with many joins/groupings.
Override first and last at the model level. This does not replicate the same functionality of the first and last methods (limit, etc)
Outside of explicitly calling order each time you use .first and .last or monkey patching Active Record, there don't seem to be good solutions available.