Extract getConfig and createWebSocketURL methods to exported functions
We're going to turn the ActionCable object into a series of
independently-importable exports so that we can remove the circular
dependencies.

We're starting with the lowest hanging fruit: `getConfig` and
`createWebSocketURL` do not reference any properties of the ActionCable
object, so they can simply be extracted out as regular functions that
we export.