Shorten unnecessarily long names for Guides
The names are only used in the Guides Index overlay and several of them are
longer than they need to be.

For example “Ruby on Rails” is mentioned over and over again in the Release
Notes section although these are obviously all Rails versions we’re listing.

There’s a Maintenance Policy category with a Maintenance Policy article which
is redundant. Policies makes much more sense as a category.

Similarly, Contributing to Ruby on Rails is both a category and an article.
“Contributions” is a better category title. “Ruby on Rails Guides Guidelines” is
a gigantic mouthful. We are *in* the Rails Guides, we don’t need to specify
what guides we’re talking about here.

As a final added advantage, all of these shortenings make the index much easier
to parse in order to find an article.

[ci skip]