Reduce string allocations in read/write_attribute
When `attr_name` is passed as a symbol, it's currently converted to a
string by `attribute_alias?`, and potentially also `attribute_alias`,
as well as by the `read_attribute`/`write_attribute` method itself.

By converting `attr_name` to a string up front, the extra allocations
related to attribute aliases can be avoided.