Tweak queue_classic and pg versions
The queue_classic gem before version 3.1.0 was locked to patch releases
of the pg gem so bump the version up from 0.3.2 - unsure how this ever
worked with 0.3.2 as QC::Setup module didn't exist until version 2.0.0.

Even with using queue_classic-3.1.0 it still has an upper bound of < 0.19.0
so we need to reduce our pg upper bound so that bundler will resolve.