Handle enums correctly when superclass ignores the column
Enums that are defined and unignored in a child class, but ignored by
the parent class, were incorrectly being treated as integers when the
record was loaded from the parent class.

This was due to the fact that when `find_by_sql` calls `instantiate`, it
passes a `column_types` hash containing the types of all columns not
defined in the schema. But when the record is loaded through the parent
class, this schema is the *parent* class's schema, which doesn't include
the enum column. As a result, when the AttributeSet is constructed, the
enum column is included both in `attribute_types` (where it's correctly
defined as an EnumType) and in `additional_types` (where it's
incorrectly defined as an Integer). The type in `additional_types`
overrides the default type derived from the schema, so
`child_instance.my_enum_attribute` returns an integer value rather than
the expected string.

To address this, we change `find_by_sql` to filter out all columns in
`possible_column_names`, which includes any ignored columns that are
unignored in a descendant class, rather than `columns_hash.keys`, which
omits all ignored columns in the current class, regardless of whether
they're unignored in a descendant.